Fix jasmine:integration Task Failure When No Specs Exist
`grunt-contrib-jasmine` fails task runs with a warning when no specs
exist. This issue was not fixed in versions supporting Jasmine 1.3,
so the plugin target is pointing at a fork with the fix.

NOTE: If/when we move to Jasmine 2, re-evaulate whether this
workaround is still necessary.

The issue was reported here:
https://github.com/gruntjs/grunt-contrib-jasmine/pull/90/files